requirements.txt
pdfminer.six==20221105
PyPDF4==1.27.0
tabula-py==2.7.0
